mdbx: 4Gb initial DB size (to make the benchmark conditions the same as for LMDB).
Several users reported that the benchmark results they
observed do not match the claims that libmdbx is slightly
faster than LMDB and in fact they see the opposite.

The reason is that the libmdbx driver was sets the dynamic
database size, which is why additional time was spent on
increasing the file.

This commit fixes this.